Static Crosscompile arm v0.9.1

+++ graphdriver/btrfs/btrfs.go
@@ -1,4 +1,4 @@
-// +build linux,amd64
+// +build linux,amd64 linux,arm
 
 package btrfs
 
@@ -30,7 +30,7 @@ func Init(home string) (graphdriver.Driver, error) {
 		return nil, err
 	}
 
-	if buf.Type != 0x9123683E {
+	if uint(buf.Type) != uint(0x9123683E) {
 		return nil, fmt.Errorf("%s is not a btrfs filesystem", rootdir)
 	}
